+++ crates/stub/src/crypto.rs
@@ -13,5 +13,6 @@
 // limitations under the License.
 
 mod ec;
+mod ecdh;
 mod ecdsa;
 mod hash;